requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
sqlparse==0.1.19
